Add StackNavigator to ProfileView
